+++ CHANGELOG.md
@@ -12,6 +12,8 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 
 ### Changes
 
+- [#2380](https://gitlab.com/meltano/meltano/-/issues/2380) Add `target_` prefix to `namespace`s of discoverable loaders target-postgres, target-snowflake, and target-sqlite
+
 ### Fixes
 
 - [#2373](https://gitlab.com/meltano/meltano/-/issues/2373) `meltano init` emits warning instead of failing when underlying filesystem doesn't support symlinks